updated unit test based on feedback, change from double to bigdecimal to resolve delta issue

+++ src/main/java/org/example/model/Item.java
@@ -1,12 +1,15 @@
 package org.example.model;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
+
 @Data
 public class Item {
     private String name;
-    private double price;
+    private BigDecimal price;
     private Integer qty;
-    public Item(String name, double price, Integer qty) {
+    public Item(String name, BigDecimal price, Integer qty) {
         this.name = name;
         this.price = price;
         this.qty = qty;
